+++ service/controller/rbac/resource/fluxauth/resource.go
@@ -8,7 +8,7 @@
 package fluxauth
 
 import (
-	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"