+++ service/controller/rbac/resource/namespaceauth/create.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/giantswarm/apiextensions/v3/pkg/label"
+	k8smetadata "github.com/giantswarm/k8smetadata/pkg/label"
 	"github.com/giantswarm/microerror"
 	rbacv1 "k8s.io/api/rbac/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -23,6 +23,10 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		return microerror.Mask(err)
 	}
 
+	if !key.HasOrganizationOrCustomerLabel(ns) {
+		return nil
+	}
+
 	// Create ClusterRole allowing 'get' access to Organization CR
 	{
 		orgReadClusterRoleName := pkgkey.OrganizationReadClusterRoleName(ns.Name)
@@ -35,7 +39,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 			ObjectMeta: metav1.ObjectMeta{
 				Name: orgReadClusterRoleName,
 				Labels: map[string]string{
-					label.ManagedBy: project.Name(),
+					k8smetadata.ManagedBy: project.Name(),
 				},
 			},
 			Rules: []rbacv1.PolicyRule{
@@ -78,7 +82,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 			ObjectMeta: metav1.ObjectMeta{
 				Name: roleBindingToCustomerGroupName,
 				Labels: map[string]string{
-					label.ManagedBy: project.Name(),
+					k8smetadata.ManagedBy: project.Name(),
 				},
 			},
 			Subjects: []rbacv1.Subject{
@@ -132,7 +136,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 			ObjectMeta: metav1.ObjectMeta{
 				Name: roleBindingToAutomationSAName,
 				Labels: map[string]string{
-					label.ManagedBy: project.Name(),
+					k8smetadata.ManagedBy: project.Name(),
 				},
 			},
 			Subjects: []rbacv1.Subject{